requirements.txt
streamlit
azure-storage-blob
psycopg2
dotenv